First week of actually (effectively) working on the project
I'm currently trying to make my main experiment work. That is, perform the spread computation on a jailbreak dataset, run the model against the fuzzer with those same examples as the sources and determine the correlation between these two. For this, I need to collect various jailbreaking datasets. Once I get the fuzzer working, I will focus on the dataset and the exacts of the experiment. I do not need to focus on these for now since I will need the fuzzer (or something similar) anyway.